requirements.txt
# Core Framework
Flask==3.0.0
python-dotenv==1.0.0

# Google Cloud & AI
google-generativeai==0.3.2
google-cloud-texttospeech==2.14.1
google-auth==2.23.4

# Audio Processing
pydub==0.25.1

# WSGI Server (Production)
gunicorn==21.2.0
gevent==23.9.1

# Optional: Code Quality & Development
# flask-cors==4.0.0  # If you need CORS support
# pytest==7.4.3  # For testing
